requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
tensorflow
